requirements.txt
Django
psycopg2-binary
django-widget-tweaks
django-betterforms
